requirements.txt
selenium==4.22.0
python-dotenv==1.0.1
#notifications
apprise==1.8.0